Implement a substrings method
Add the substrings method which scans a string for each word in
a dictionary and assings the number of each word in a hash with
key corresponding to the word.

This method follows every project's requirements, it doesn't
differentiate capital letters from the string, but it does for
the dictionary.